@@ -1,5 +1,6 @@
-import ReactDOM from 'react-dom';
-import App from './App';
-import './index.css';
+import ReactDOM from "react-dom/client";
+import App from "./App";
+import "./index.css";
 
-ReactDOM.render(<App/>, document.querySelector("#root"))
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(<App />);
