@@ -23,24 +23,24 @@ const About = () => {
             <article className="about__card">
               <FaAward className="about__icon" />
               <h5>Experience</h5>
-              <small>3.5+ Years</small>
+              <small>4.5+ Years</small>
             </article>
             <article className="about__card">
               <FiUsers className="about__icon" />
               <h5>Clients</h5>
-              <small>6+ in Europe</small>
+              <small>11+ in Europe</small>
             </article>
             <article className="about__card">
               <VscFolderLibrary className="about__icon" />
               <h5>Projects</h5>
-              <small>13+ Projects</small>
+              <small>15+ Projects</small>
             </article>
           </div>
 
           <p>
-            Frontend developer who helps people materialize their dream into a
+            I'm Frontend developer who helps people materialize their dream into a
             beautiful, hands-on prototype. Creative, insightful and passionate
-            about writing, videography and voley.
+            about writing, videography and volley.
           </p>
 
           <a href="#contact" className="btn btn-primary">
