@@ -1,13 +1,19 @@
 import React from 'react'
-import CV from '../../assets/GiuliaIvanResume.pdf'
+// import CV from '../../assets/GiuliaIvanResume.pdf'
+// import CV from "../../assets/GiuliaIvanResume1.pdf";
+import CV from "../../assets/GiuliaIvanResume2.pdf";
 
 const CTA = () => {
   return (
-    <div className='cta'>
-        <a href={CV} download className='btn'>Download CV</a>
-        <a href="#contact" className='btn btn-primary'>Let's Talk</a>
+    <div className="cta">
+      <a href={CV} download className="btn btn-primary">
+        Download CV
+      </a>
+      <a href="#contact" className="btn">
+        Let's Talk
+      </a>
     </div>
-  )
+  );
 }
 
 export default CTA
